Displays greeting message everytime a attendee checks in
Number of  attendees are tracked with a visual counter

Team attendance is tracked with a visual counter below

Progress bar fills up by percentage as long as attendees check in. Reaches and stops at 100% oncew attendee reaches max: 50.

When attendance reaches 50 pop up message shows a winner message of which team has the most attendees

Added a section to track attendee names and the team they are from that displays visually.

Attendance trackers,  progress bar, and names of attendees are saved even though browser is refreshed or closed